# [Bridges] improve function getter of SplitHyperRectangleBridge