@odow 2025-03-05
This needs some better tests. It doesn't account for a few things